fix: fix failing tests and restructure files
Move all hooks into `hooks` folder; Extract data-fetching hooks
to useFetchPipelineSettingsData; rename useFetchPipelineSettings to
usePipelineSettings
