+++ services/orchest-webserver/client/src/types.d.ts
@@ -276,7 +276,7 @@ export type Step = {
     name?: string;
   };
   meta_data: { hidden: boolean; position: [number, number] };
-  parameters: Json;
+  parameters: Record<string, Json>;
 };
 
 export type MouseTracker = {
